Initial commit of mint pipeline
* project_init.sh with a project name for an argument initializes the
directory structure for the project.
* project_create_runs.R uses an annotation file (described extensively
at the top of the script) to determine how the pipeline should be
constructed. Hybrid and pure pulldown experiments are supported.
Matched and shared input samples are supported. Simple, sample (hybrid
only), and comparison classifications are supported.
TO DO
1. Test on sample data / different experimental setups, and workout the
bugs.
2. Develop sample classification for pure pulldown setup.
3. Develop the visualization module as the final part of the pipeline
once all analysis steps are complete.