0.1.1 Pipeline fixes
1. In scripts/process_bisulfite_align.sh, there was an unexpected
problem where the bismark_methylation_extractor with —bedGraph
mistakenly interpreted path/to/CpG* files used to create the bedGraph
as not beginning with CpG because of the path/to/. To fix this, we cd
into ${ANALAYSISDIR}/bismark_extractor_calls before running the
methylation extractor. We correspondingly change the Bismark bedGraph
file paths. This seems to solve the problem.

2. PePr does not have an output directory parameter, so in
scripts/process_pulldown_comparison.sh, to ensure results go to
${ANALYSISDIR}/pepr_peaks/ we cd into that directory before running
PePr.

3. Removed —remove_artefacts parameter because an error is thrown in
conjunction with —threshold. Yanxiao has not been able to figure out
why this happens.

4. In scripts/project_create_runs.R we neglected to add the path to the
BAMS from bowtie2 which are used as PePr input. This resulted in the
files not being found and the analysis failing. A similar problem was
going to happen in the methylSig run because the
bismark_methylation_extractor files referenced did not have /path/to/.