0.3.5 Fixing relative paths
* In scripts/process_bisulfite_align.sh and
scripts/process_pulldown_comparison.sh we deviate from the norm of
executing within the project root because of issues with Bismark and
PePr, respectively. The track hub path should therefore begin with
../summary/… instead of ./analysis/…

* gzip the Bismark bedGraph file after creating the browser track.

* In creation of the script running PePr, the bowtie2_bams are actually
in ../bowtie2_bams relative to the execution context of
analysis/pepr_peaks.

* In creation of MACS2 scripts, only need to give the IDs, not the full
path, because that’s taken care of in scripts/process_pulldown_sample.sh